Edits from Geroge Timson:
Sam, I just changed the file named dba_frm1.htm.  There was a
longstanding error that pointed  'CREIXN^DDMOD' to the wrong page. I'll
have a couple of other changes to make to the programmer manual.

pm_inde2.htm is the other file I have changed at hardhats.org.  Here the
error was an error of omission.  There were 4 APIs that needed to be
included in the comprehensive "API Index".